requirements.txt
# OpenAI API
openai>=1.0.0

# HuggingFace transformers and PyTorch
transformers>=4.21.0
torch>=2.0.0
tokenizers>=0.13.0

# NLTK for sentence tokenization
nltk>=3.8

# Environment management
python-dotenv>=1.0.0

# Utilities
termcolor>=2.3.0
numpy>=1.24.0

# Local vector operations
scikit-learn>=1.3.0

# Optional: for additional text processing
# chonkie  # Add if needed for alternative chunking